+++ Source/Game/Spells/SpellEffects.cs
@@ -4129,7 +4129,7 @@ namespace Game.Spells
                 //GO is always friendly to it's creator, get range for friends
                 float min_dis = m_spellInfo.GetMinRange(true);
                 float max_dis = m_spellInfo.GetMaxRange(true);
-                float dis = (float)RandomHelper.NextDouble() * (max_dis - min_dis) + min_dis;
+                float dis = RandomHelper.NextSingle() * (max_dis - min_dis) + min_dis;
 
                 unitCaster.GetClosePoint(out fx, out fy, out fz, SharedConst.DefaultPlayerBoundingRadius, dis);
                 fo = unitCaster.GetOrientation();
